requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
redis>=5.0.0
channels>=4.0.0
channels-redis>=4.1.0
python-chess>=1.999
django-environ>=0.11.2
Pillow>=10.1.0
whitenoise>=6.6.0
daphne>=4.0.0
gunicorn>=21.2.0
